split physics into modules
