+++ chariot-server/src/physics/bounding_box.rs
@@ -1,0 +1,2 @@
+
+pub type BoundingBoxDimensions = [[f64; 2]; 3];
