+++ chariot-server/src/physics/player_entity.rs
@@ -1,10 +1,16 @@
 use chariot_core::entity_location::EntityLocation;
 use chariot_core::lap_info::LapInformation;
-use chariot_core::physics_changes::PhysicsChange;
-use chariot_core::player_inputs::PlayerInputs;
+use chariot_core::physics_changes::{PhysicsChange, PhysicsChangeType};
+use chariot_core::player_inputs::{EngineStatus, PlayerInputs, RotationStatus};
 use glam::DVec3;
+use chariot_core::GLOBAL_CONFIG;
 
-pub type BoundingBoxDimensions = [[f64; 2]; 3];
+use crate::physics::bounding_box::BoundingBoxDimensions;
+use crate::physics::trigger_entity::TriggerEntity;
+
+fn get_height_at_coordinates(_x: f64, _z: f64) -> f64 {
+    return 0.0;
+}
 
 pub struct PlayerEntity {
     pub velocity: DVec3,
@@ -20,4 +26,205 @@ pub struct PlayerEntity {
     pub physics_changes: Vec<PhysicsChange>,
 
     pub lap_info: LapInformation,
+}
+
+impl PlayerEntity {
+    pub fn set_upward_direction_from_bounding_box(&mut self) {
+        let [min_x, max_x] = self.bounding_box[0];
+        let [min_z, max_z] = self.bounding_box[2];
+
+        let lower_left_corner = DVec3::new(min_x, get_height_at_coordinates(min_x, min_z), min_z);
+        let lower_right_corner = DVec3::new(max_x, get_height_at_coordinates(max_x, min_z), min_z);
+        let upper_left_corner = DVec3::new(min_x, get_height_at_coordinates(min_x, max_z), max_z);
+        let upper_right_corner = DVec3::new(max_x, get_height_at_coordinates(max_x, max_z), max_z);
+
+        let diagonal_1 = lower_right_corner - upper_left_corner;
+        let diagonal_2 = upper_right_corner - lower_left_corner;
+
+        self.entity_location.unit_upward_direction = diagonal_2.cross(diagonal_1).normalize();
+    }
+
+    /* Given a set of physical properties, compute and return what next tick's
+     * physics properties will be for that object */
+    pub fn do_physics_step(
+        &self,
+        time_step: f64,
+        potential_colliders: Vec<&PlayerEntity>,
+        potential_triggers: Vec<Box<&dyn TriggerEntity>>,
+    ) -> PlayerEntity {
+        let self_forces = self.sum_of_self_forces();
+        let acceleration = self_forces / self.mass;
+
+        let angular_velocity: f64 = match self.player_inputs.rotation_status {
+            RotationStatus::InSpinClockwise => f64::min(
+                GLOBAL_CONFIG.max_car_spin,
+                self.angular_velocity + GLOBAL_CONFIG.car_spin,
+            ),
+            RotationStatus::InSpinCounterclockwise => f64::max(
+                -GLOBAL_CONFIG.max_car_spin,
+                self.angular_velocity - GLOBAL_CONFIG.car_spin,
+            ),
+            RotationStatus::NotInSpin => {
+                self.angular_velocity * GLOBAL_CONFIG.rotation_reduction_coefficient
+            }
+        };
+
+        let rotation_matrix = glam::Mat3::from_axis_angle(
+            self.entity_location.unit_upward_direction.as_vec3(),
+            -1.0 * angular_velocity as f32,
+        );
+
+        let new_steer_direction = rotation_matrix
+            .mul_vec3(self.entity_location.unit_steer_direction.as_vec3())
+            .normalize()
+            .as_dvec3();
+
+        let mut delta_velocity = acceleration * time_step;
+
+        for collider in potential_colliders.iter() {
+            delta_velocity += self.delta_v_from_collision_with_player(collider);
+        }
+
+        let mut new_velocity = self.velocity + delta_velocity;
+        if new_velocity.length() > GLOBAL_CONFIG.max_car_speed {
+            new_velocity = new_velocity.normalize() * GLOBAL_CONFIG.max_car_speed;
+        }
+
+        let mut new_player = PlayerEntity {
+            player_inputs: PlayerInputs {
+                engine_status: self.player_inputs.engine_status,
+                rotation_status: self.player_inputs.rotation_status,
+            },
+
+            entity_location: EntityLocation {
+                position: self.entity_location.position + self.velocity * time_step,
+                unit_steer_direction: new_steer_direction,
+                unit_upward_direction: self.entity_location.unit_upward_direction,
+            },
+
+            velocity: new_velocity,
+            angular_velocity: angular_velocity,
+            mass: self.mass,
+            size: self.size,
+            bounding_box: self.bounding_box,
+            physics_changes: self.physics_changes.clone(),
+            lap_info: self.lap_info,
+        };
+
+        new_player.apply_physics_changes();
+
+        for b in potential_triggers.iter() {
+            if b.check_bounding_box_collisions(&new_player) {
+                b.trigger(&mut new_player);
+            }
+        }
+
+        return new_player;
+    }
+
+    fn is_aerial(&self) -> bool {
+        return self.entity_location.position[1]
+            > self.size[1]
+            + get_height_at_coordinates(
+            self.entity_location.position[0],
+            self.entity_location.position[2],
+        );
+    }
+
+    fn sum_of_self_forces(&self) -> DVec3 {
+        let air_forces = self.gravitational_force_on_object()
+            + self.player_applied_force_on_object()
+            + self.air_resistance_force_on_object();
+
+        return if self.is_aerial() {
+            air_forces
+        } else {
+            air_forces + self.normal_force_on_object() + self.rolling_resistance_force_on_object()
+        };
+    }
+
+    fn gravitational_force_on_object(&self) -> DVec3 {
+        return DVec3::new(0.0, -1.0, 0.0) * self.mass * GLOBAL_CONFIG.gravity_coefficient;
+    }
+
+    fn normal_force_on_object(&self) -> DVec3 {
+        return self.entity_location.unit_upward_direction * self.mass;
+    }
+
+    // Includes two player-applied forces: accelerator and brake.
+    fn player_applied_force_on_object(&self) -> DVec3 {
+        match self.player_inputs.engine_status {
+            // The accelerator always applies forces in the steer direction; since
+            // rotation is free, this is the intuitive direction. Braking, however,
+            // is directionless, so the force of braking applies in a different
+            // direction: specifically, it acts against whatever the current
+            // direction of travel is. (which is not the steering direction!)
+            EngineStatus::Accelerating => {
+                return self.entity_location.unit_steer_direction
+                    * self.mass
+                    * GLOBAL_CONFIG.car_accelerator
+            }
+            // divide velocity by its magnitude to have a unit vector pointing
+            // towards current heading, then apply the force in the reverse direction
+            EngineStatus::Braking => {
+                return self.velocity / self.velocity.length()
+                    * -1.0
+                    * self.mass
+                    * GLOBAL_CONFIG.car_brake
+            }
+            // And there is no player-applied force when not accelerating or braking
+            EngineStatus::Neutral => return DVec3::new(0.0, 0.0, 0.0),
+        }
+    }
+
+    // Equations for modelling air resistance and rolling resistance come from
+    // https://asawicki.info/Mirror/Car%20Physics%20for%20Games/Car%20Physics%20for%20Games.html
+    fn air_resistance_force_on_object(&self) -> DVec3 {
+        // air resistance is proportional to the square of velocity
+        return self.velocity
+            * self.mass
+            * -1.0
+            * GLOBAL_CONFIG.drag_coefficient
+            * self.velocity.length();
+    }
+
+    fn rolling_resistance_force_on_object(&self) -> DVec3 {
+        return self.velocity * self.mass * -1.0 * GLOBAL_CONFIG.rolling_resistance_coefficient;
+    }
+
+    fn apply_physics_changes(&mut self) {
+        for change in &self.physics_changes {
+            match change.change_type {
+                PhysicsChangeType::IAmSpeed => {
+                    let flat_speed_increase = 30.0;
+                    self.velocity = self.velocity * (self.velocity.length() + flat_speed_increase);
+                }
+                PhysicsChangeType::NoTurningRight => {
+                    if matches!(
+                        self.player_inputs.rotation_status,
+                        RotationStatus::InSpinClockwise
+                    ) {
+                        self.player_inputs.rotation_status = RotationStatus::NotInSpin;
+                        self.angular_velocity -= GLOBAL_CONFIG.car_spin;
+                    }
+                }
+                PhysicsChangeType::ShoppingCart => {
+                    self.angular_velocity += GLOBAL_CONFIG.car_spin / 2.0;
+                }
+                PhysicsChangeType::InSpainButTheAIsSilent => {
+                    match self.player_inputs.rotation_status {
+                        RotationStatus::InSpinClockwise => {}
+                        RotationStatus::NotInSpin => {
+                            self.player_inputs.rotation_status = RotationStatus::InSpinClockwise;
+                            self.angular_velocity += GLOBAL_CONFIG.car_spin;
+                        }
+                        RotationStatus::InSpinCounterclockwise => {
+                            self.player_inputs.rotation_status = RotationStatus::InSpinClockwise;
+                            self.angular_velocity += 2.0 * GLOBAL_CONFIG.car_spin;
+                        }
+                    }
+                }
+            }
+        }
+    }
 }
